requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0

langgraph>=0.0.45
langchain-core>=0.1.40

google-generativeai>=0.5.0
typing_extensions>=4.7.0
